requirements.txt
efficientnet_pytorch==0.6.3
einops==0.4.1
imageio==2.18.0
imbalanced_learn==0.9.0
imblearn==0.0
kornia==0.6.6
matplotlib==3.2.2
monai==0.8.1
numpy==1.21.6
opencv_python==4.5.5.64
pandas==1.3.5
Pillow==9.2.0
pydicom==2.3.0
pytorch_lightning==1.6.3
scikit_learn==1.0.2
seaborn==0.11.2
SimpleITK==2.2.0
torch==1.11.0
torchmetrics==0.8.1
torchvision==0.12.0
tqdm==4.64.0
